# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for the Voxlab public speaking educational platform",
        "title": "Voxlab API",
        "termsOfService": "http://voxlab.com/terms",
        "contact": {
            "name": "API Support",
            "email": "support@voxlab.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/progress": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns ALL progress records for the authenticated user (every lesson they've started or completed).\nEach record shows: status (in_progress/completed), xp_earned, and timestamps.\nUse this on the frontend to determine which lessons are completed and show user progress bars.\n\n🔒 Requires JWT token (Authorization: Bearer \u003ctoken\u003e)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Progress"
                ],
                "summary": "Get my learning progress",
                "responses": {
                    "200": {
                        "description": "Registros de progreso del usuario",
                        "schema": {
                            "$ref": "#/definitions/resources.ListProgressResponse"
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado o inválido",
                        "schema": {
                            "$ref": "#/definitions/resources.UnauthorizedError"
                        }
                    },
                    "500": {
                        "description": "Error al obtener el progreso",
                        "schema": {
                            "$ref": "#/definitions/resources.InternalServerError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Marks a lesson as completed for the authenticated user. This endpoint:\n1. Sets the progress status to \"completed\"\n2. Adds XP to the user's total (calculated from exercises + score)\n3. Stores the completion timestamp\n\nIf the lesson was already completed before, it UPDATES the existing record\n(the final score replaces the previous one).\n\n🔒 Requires JWT token (Authorization: Bearer \u003ctoken\u003e)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Progress"
                ],
                "summary": "Complete a lesson",
                "parameters": [
                    {
                        "description": "Lesson completion data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.completeLessonRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Progreso actualizado — lección completada",
                        "schema": {
                            "$ref": "#/definitions/resources.CompleteProgressResponse"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos — lesson_id requerido",
                        "schema": {
                            "$ref": "#/definitions/resources.BadRequestError"
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado o inválido",
                        "schema": {
                            "$ref": "#/definitions/resources.UnauthorizedError"
                        }
                    },
                    "404": {
                        "description": "Lección no encontrada",
                        "schema": {
                            "$ref": "#/definitions/resources.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Error al completar la lección",
                        "schema": {
                            "$ref": "#/definitions/resources.InternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.completeLessonRequest": {
            "type": "object",
            "properties": {
                "lesson_id": {
                    "type": "integer",
                    "example": 1
                },
                "score": {
                    "type": "integer",
                    "example": 85
                }
            }
        },
...
        "models.UserProgress": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "lesson_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "xp_earned": {
                    "type": "integer"
                }
            }
        },
...
        "resources.BadRequestError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Solicitud inválida — verifique los campos enviados"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "resources.CompleteProgressResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.UserProgress"
                },
                "message": {
                    "type": "string",
                    "example": "Operación realizada con éxito"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "resources.InternalServerError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error interno del servidor"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "resources.ListProgressResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserProgress"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Operación realizada con éxito"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "resources.NotFoundError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Recurso no encontrado"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "resources.UnauthorizedError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Token de autenticación no proporcionado o inválido"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}